Moved test code into src/tests/utils as a Boost unit test Renamed calibrated_salinity.h to derived_salinity.h
