+++ src/bin/fusion/fusion.cpp
@@ -44,7 +44,7 @@
 #include "jaiabot/messages/modem_message_extensions.pb.h"
 #include "jaiabot/messages/sensor/pressure_temperature.pb.h"
 #include "jaiabot/messages/sensor/salinity.pb.h"
-#include "jaiabot/utils/calibrated_salinity.h"
+#include "jaiabot/utils/derived_salinity.h"
 #include "wmm/WMM.h"
 #include <cmath>
 #include <math.h>
