+++ src/lib/utils/derived_salinity.h
@@ -1,6 +1,3 @@
-#include <assert.h>
-#include <iomanip>
-#include <iostream>
 #include <math.h>
 
 double calculate_derived_salinity(const double measured_conductivity, const double temperature,
@@ -68,33 +65,4 @@ double calculate_derived_salinity(const double measured_conductivity, const doub
                      (a5 * pow(Rt, 2.5)) + dS;
 
     return S;
-}
-
-bool test_calculate_derived_salinity()
-{
-    struct TestData
-    {
-        double R, t, p, S;
-    };
-
-    const TestData tests[] = {
-        {1.0, 15.0, 0.0, 35.0}, {1.2, 20.0, 2000.0, 37.245628}, {0.65, 5.0, 1500.0, 27.995347}};
-
-    const double standard_conductivity = 42914;
-
-    for (auto test : tests)
-    {
-        const double measured_conductivity = test.R * standard_conductivity;
-        const double S = calculate_derived_salinity(measured_conductivity, test.t, test.p);
-
-        if (float(S) != float(test.S))
-        {
-            std::cout << std::setprecision(10) << "Test failed: R = " << test.R
-                      << ", t = " << test.t << ", p = " << test.p << ", S = " << test.S
-                      << ", got S = " << S << std::endl;
-            return false;
-        }
-    }
-
-    return true;
 }
