+++ src/test/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_subdirectory(comms)
 add_subdirectory(sensor)
+add_subdirectory(utils)
