+++ src/test/utils/test.cpp
@@ -1,0 +1,35 @@
+#define BOOST_TEST_MODULE jaiabot_test_utils
+#include "jaiabot/utils/derived_salinity.h"
+#include <boost/test/included/unit_test.hpp>
+
+namespace jaiabot
+{
+namespace utils
+{
+BOOST_AUTO_TEST_SUITE(utils_tests)
+
+BOOST_AUTO_TEST_CASE(test_conductivity)
+{
+    struct TestData
+    {
+        double R, t, p, S;
+    };
+
+    const TestData tests[] = {
+        {1.0, 15.0, 0.0, 35.0}, {1.2, 20.0, 2000.0, 37.245628}, {0.65, 5.0, 1500.0, 27.995347}};
+
+    const double standard_conductivity = 42914;
+
+    for (auto test : tests)
+    {
+        const double measured_conductivity = test.R * standard_conductivity;
+        const double S = calculate_derived_salinity(measured_conductivity, test.t, test.p);
+
+        BOOST_CHECK_CLOSE(S, test.S, 0.00001);
+    }
+}
+
+BOOST_AUTO_TEST_SUITE_END()
+
+} // namespace utils
+} // namespace jaiabot
